[types] Merge two implementations of NewBlockEvent
