+++ types/src/block_metadata.rs
@@ -132,56 +132,3 @@ impl MoveStructType for BlockResource {
 }
 
 impl MoveResource for BlockResource {}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct NewBlockEvent {
-    epoch: u64,
-    round: u64,
-    previous_block_votes: Vec<bool>,
-    proposer: AccountAddress,
-    failed_proposer_indices: Vec<u64>,
-    timestamp: u64,
-}
-
-impl NewBlockEvent {
-    pub fn new(
-        epoch: u64,
-        round: u64,
-        previous_block_votes: Vec<bool>,
-        proposer: AccountAddress,
-        failed_proposer_indices: Vec<u64>,
-        timestamp: u64,
-    ) -> Self {
-        Self {
-            epoch,
-            round,
-            previous_block_votes,
-            proposer,
-            failed_proposer_indices,
-            timestamp,
-        }
-    }
-
-    pub fn epoch(&self) -> u64 {
-        self.epoch
-    }
-
-    pub fn round(&self) -> u64 {
-        self.round
-    }
-
-    pub fn previous_block_votes(&self) -> &Vec<bool> {
-        &self.previous_block_votes
-    }
-
-    pub fn proposer(&self) -> AccountAddress {
-        self.proposer
-    }
-
-    /// The list of indices in the validators list,
-    /// of consecutive proposers from the immediately preceeding
-    /// rounds that didn't produce a successful block
-    pub fn failed_proposer_indices(&self) -> &Vec<u64> {
-        &self.failed_proposer_indices
-    }
-}
